+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # tnsping2
 
+## [1.1.0 - 2024-04-27]
+- added echo Server and Client
+- update dependencies
+- reduce complexity for ping tcp
+
 ## [v1.0.1 - 2024-04-13]
 - add MTR_BIN environment variable to set mtr binary path
 - update pipeline to install dependencies